+++ src/store.js
@@ -63,7 +63,7 @@ Project*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.store = '2021-November-10';
+modules.store = '2021-November-11';
 
 // XML_Serializer ///////////////////////////////////////////////////////
 /*
@@ -380,6 +380,9 @@ SnapSerializer.prototype.loadScene = function (xmlNode, remixID) {
     }
     scene.unifiedPalette = model.scene.attributes.palette === 'single';
     scene.showCategories = model.scene.attributes.categories !== 'false';
+    scene.disableClickToRun = model.scene.attributes.clickrun === 'false';
+    scene.penColorModel = model.scene.attributes.colormodel === 'hsv' ?
+        'hsv' : 'hsl';
     model.notes = model.scene.childNamed('notes');
     if (model.notes) {
         scene.notes = model.notes.contents;
@@ -1735,7 +1738,7 @@ Scene.prototype.toXML = function (serializer) {
     }
 
     xml = serializer.format(
-        '<scene name="@"%%>' +
+        '<scene name="@"%%%%>' +
             '<notes>$</notes>' +
             '%' +
             '<hidden>$</hidden>' +
@@ -1749,6 +1752,8 @@ Scene.prototype.toXML = function (serializer) {
         this.unifiedPalette ? ' palette="single"' : '',
         this.unifiedPalette && !this.showCategories ?
             ' categories="false"' : '',
+        this.disableClickToRun ? ' clickrun="false"' : '',
+        this.penColorModel === 'hsv' ? ' colormodel="hsv"' : '',
         this.notes || '',
         serializer.paletteToXML(this.customCategories),
         Object.keys(this.hiddenPrimitives).reduce(